Migrated many tests from test_atomgroup into smaller test files
