+++ testsuite/MDAnalysisTests/core/test_universe.py
@@ -37,10 +37,14 @@ from MDAnalysisTests.datafiles import (
     PSF, DCD,
     PSF_BAD,
     PDB_small,
+    GRO, TRR,
+    two_water_gro, two_water_gro_nonames,
+    TRZ, TRZ_psf,
 )
 from MDAnalysisTests import parser_not_found
 
 import MDAnalysis as mda
+import MDAnalysis.coordinates
 from MDAnalysis.topology.base import TopologyReader
 
 
@@ -177,3 +181,198 @@ class TestUniverse(object):
         box = np.array([10, 11, 12, 90, 90, 90])
         u.dimensions = np.array([10, 11, 12, 90, 90, 90])
         assert_allclose(u.dimensions, box)
+
+
+class TestGuessBonds(object):
+    """Test the AtomGroup methed guess_bonds
+
+    This needs to be done both from Universe creation (via kwarg) and AtomGroup
+
+    It needs to:
+     - work if all atoms are in vdwradii table
+     - fail properly if not
+     - work again if vdwradii are passed.
+    """
+    def setUp(self):
+        self.vdw = {'A':1.05, 'B':0.4}
+
+    def tearDown(self):
+        del self.vdw
+
+    def _check_universe(self, u):
+        """Verify that the Universe is created correctly"""
+        assert_equal(len(u.bonds), 4)
+        assert_equal(len(u.angles), 2)
+        assert_equal(len(u.dihedrals), 0)
+        assert_equal(len(u.atoms[0].bonds), 2)
+        assert_equal(len(u.atoms[1].bonds), 1)
+        assert_equal(len(u.atoms[2].bonds), 1)
+        assert_equal(len(u.atoms[3].bonds), 2)
+        assert_equal(len(u.atoms[4].bonds), 1)
+        assert_equal(len(u.atoms[5].bonds), 1)
+        assert_('guess_bonds' in u.kwargs)
+
+    def test_universe_guess_bonds(self):
+        """Test that making a Universe with guess_bonds works"""
+        u = mda.Universe(two_water_gro, guess_bonds=True)
+        self._check_universe(u)
+        assert_(u.kwargs['guess_bonds'] is True)
+
+    def test_universe_guess_bonds_no_vdwradii(self):
+        """Make a Universe that has atoms with unknown vdwradii."""
+        assert_raises(ValueError, mda.Universe, two_water_gro_nonames, guess_bonds=True)
+
+    def test_universe_guess_bonds_with_vdwradii(self):
+        """Unknown atom types, but with vdw radii here to save the day"""
+        u = mda.Universe(two_water_gro_nonames, guess_bonds=True,
+                                vdwradii=self.vdw)
+        self._check_universe(u)
+        assert_(u.kwargs['guess_bonds'] is True)
+        assert_equal(self.vdw, u.kwargs['vdwradii'])
+
+    def test_universe_guess_bonds_off(self):
+        u = mda.Universe(two_water_gro_nonames, guess_bonds=False)
+
+        for attr in ('bonds', 'angles', 'dihedrals'):
+            assert_(not hasattr(u, attr))
+        assert_(u.kwargs['guess_bonds'] is False)
+
+    def _check_atomgroup(self, ag, u):
+        """Verify that the AtomGroup made bonds correctly,
+        and that the Universe got all this info
+        """
+        assert_equal(len(ag.bonds), 2)
+        assert_equal(len(ag.angles), 1)
+        assert_equal(len(ag.dihedrals), 0)
+        assert_equal(len(u.bonds), 2)
+        assert_equal(len(u.angles), 1)
+        assert_equal(len(u.dihedrals), 0)
+        assert_equal(len(u.atoms[0].bonds), 2)
+        assert_equal(len(u.atoms[1].bonds), 1)
+        assert_equal(len(u.atoms[2].bonds), 1)
+        assert_equal(len(u.atoms[3].bonds), 0)
+        assert_equal(len(u.atoms[4].bonds), 0)
+        assert_equal(len(u.atoms[5].bonds), 0)
+
+    def test_atomgroup_guess_bonds(self):
+        """Test an atomgroup doing guess bonds"""
+        u = mda.Universe(two_water_gro)
+
+        ag = u.atoms[:3]
+        ag.guess_bonds()
+        self._check_atomgroup(ag, u)
+
+    def test_atomgroup_guess_bonds_no_vdwradii(self):
+        u = mda.Universe(two_water_gro_nonames)
+
+        ag = u.atoms[:3]
+        assert_raises(ValueError, ag.guess_bonds)
+
+    def test_atomgroup_guess_bonds_with_vdwradii(self):
+        u = mda.Universe(two_water_gro_nonames)
+
+        ag = u.atoms[:3]
+        ag.guess_bonds(vdwradii=self.vdw)
+        self._check_atomgroup(ag, u)
+
+
+class TestInMemoryUniverse(object):
+    @staticmethod
+    @dec.skipif(parser_not_found('DCD'),
+               'DCD parser not available. Are you using python 3?')
+    def test_reader_w_timeseries():
+        universe = mda.Universe(PSF, DCD, in_memory=True)
+        assert_equal(universe.trajectory.timeseries(universe.atoms).shape,
+                     (3341, 98, 3),
+                     err_msg="Unexpected shape of trajectory timeseries")
+
+    @staticmethod
+    def test_reader_wo_timeseries():
+        universe = mda.Universe(GRO, TRR, in_memory=True)
+        assert_equal(universe.trajectory.timeseries(universe.atoms).shape,
+                     (47681, 10, 3),
+                     err_msg="Unexpected shape of trajectory timeseries")
+
+    @staticmethod
+    @dec.skipif(parser_not_found('DCD'),
+               'DCD parser not available. Are you using python 3?')
+    def test_reader_w_timeseries_frame_interval():
+        universe = mda.Universe(PSF, DCD, in_memory=True,
+                                       in_memory_frame_interval=10)
+        assert_equal(universe.trajectory.timeseries(universe.atoms).shape,
+                     (3341, 10, 3),
+                     err_msg="Unexpected shape of trajectory timeseries")
+
+    @staticmethod
+    def test_reader_wo_timeseries_frame_interval():
+        universe = mda.Universe(GRO, TRR, in_memory=True,
+                                       in_memory_frame_interval=3)
+        assert_equal(universe.trajectory.timeseries(universe.atoms).shape,
+                     (47681, 4, 3),
+                     err_msg="Unexpected shape of trajectory timeseries")
+
+    @staticmethod
+    @dec.skipif(parser_not_found('DCD'),
+                'DCD parser not available. Are you using python 3?')
+    def test_existing_universe():
+        universe = mda.Universe(PDB_small, DCD)
+        universe.transfer_to_memory()
+        assert_equal(universe.trajectory.timeseries(universe.atoms).shape,
+                     (3341, 98, 3),
+                     err_msg="Unexpected shape of trajectory timeseries")
+
+    @staticmethod
+    @dec.skipif(parser_not_found('DCD'),
+                'DCD parser not available. Are you using python 3?')
+    def test_frame_interval_convention():
+        universe1 = mda.Universe(PSF, DCD)
+        array1 = universe1.trajectory.timeseries(skip=10)
+        universe2 = mda.Universe(PSF, DCD, in_memory=True,
+                                        in_memory_frame_interval=10)
+        array2 = universe2.trajectory.timeseries()
+        assert_equal(array1, array2,
+                     err_msg="Unexpected differences between arrays.")
+
+
+class TestCustomReaders(object):
+    """
+    Can pass a reader as kwarg on Universe creation
+    """
+    @dec.skipif(parser_not_found('TRZ'),
+                'TRZ parser not available. Are you using python 3?')
+    def test_custom_reader(self):
+        # check that reader passing works
+        u = mda.Universe(TRZ_psf, TRZ, format=MDAnalysis.coordinates.TRZ.TRZReader)
+        assert_equal(len(u.atoms), 8184)
+
+    def test_custom_reader_singleframe(self):
+        T = MDAnalysis.topology.GROParser.GROParser
+        R = MDAnalysis.coordinates.GRO.GROReader
+        u = mda.Universe(two_water_gro, two_water_gro,
+                                topology_format=T, format=R)
+        assert_equal(len(u.atoms), 6)
+
+    def test_custom_reader_singleframe_2(self):
+        # Same as before, but only one argument to Universe
+        T = MDAnalysis.topology.GROParser.GROParser
+        R = MDAnalysis.coordinates.GRO.GROReader
+        u = mda.Universe(two_water_gro,
+                                topology_format=T, format=R)
+        assert_equal(len(u.atoms), 6)
+
+    @dec.skipif(parser_not_found('TRZ'),
+                'TRZ parser not available. Are you using python 3?')
+    def test_custom_parser(self):
+        # topology reader passing works
+        u = mda.Universe(TRZ_psf, TRZ, topology_format=MDAnalysis.topology.PSFParser.PSFParser)
+        assert_equal(len(u.atoms), 8184)
+
+    @dec.skipif(parser_not_found('TRZ'),
+                'TRZ parser not available. Are you using python 3?')
+    def test_custom_both(self):
+        # use custom for both
+        u = mda.Universe(TRZ_psf, TRZ, format=MDAnalysis.coordinates.TRZ.TRZReader,
+                         topology_format=MDAnalysis.topology.PSFParser.PSFParser)
+        assert_equal(len(u.atoms), 8184)
+
+
